requirements.txt
# DisruptionShield Coordinator - Requirements
# Install with: pip install -r requirements.txt

# Core Framework
chainlit==1.3.2
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# LLM providers
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
aiohttp==3.11.11

# Date/Time handling
python-dateutil==2.9.0

# Async support
anyio==4.7.0
